% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{ClusterFeatures}
\alias{ClusterFeatures}
\title{ClusterFeatures}
\usage{
ClusterFeatures(group)
}
\arguments{
\item{group}{Data frame with a the angular coordinates (ra and dec in radians), the redshift, the color (g-r), the apparent magnitude (r) and the id of the group for each galaxy. The columns must be named 'ra', 'dec' and 'z', 'color', 'mag' and 'id'.}
}
\value{
Data frame with the features of the galaxy cluster.
}
\description{
This function estimates the features for a galaxy cluster.
}
\examples{
ClusterFeatures(group)
}
