% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{DresslerShectmanGalaxy}
\alias{DresslerShectmanGalaxy}
\title{DresslerShectmanGalaxy}
\usage{
DresslerShectmanGalaxy(vr, r)
}
\arguments{
\item{vr}{Vector of velocities of all the galaxies in the cluster.}

\item{r}{Indixes of the neighbour galaxies to be taking into account in the delta estimation.}
}
\value{
The delta value of the Dressler-Shectman test corresponding to the galaxy.
}
\description{
This function returns the value of the Dressler-Shectman statistic for an individual galaxy with ngal = sqrt(n).
}
\examples{
DresslerShectmanGalaxy(vr, r)
}
