% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{DresslerShectmanIndividual}
\alias{DresslerShectmanIndividual}
\title{DresslerShectmanIndividual}
\usage{
DresslerShectmanIndividual(group)
}
\arguments{
\item{group}{data frame with a the angular coordinates (ra and dec in radians) and the redshift. The columns must be named 'ra', 'dec' and 'z'.}
}
\value{
The delta value of the Dressler-Shectman test corresponding to each galaxy of a galaxy cluster.
}
\description{
This function returns the delta value of the Dressler-Shectman statistic each galaxy of a galaxy cluster with ngal = 10.
}
\examples{
DresslerShectmanIndividual(group)
}
