% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{DresslerShectmanIter}
\alias{DresslerShectmanIter}
\title{DresslerShectmanIter}
\usage{
DresslerShectmanIter(cat)
}
\arguments{
\item{cat}{data frame with a the angular coordinates (ra and dec in radians) and the redshift. The columns must be named 'ra', 'dec' and 'z'.}
}
\value{
Number of iterations untill convergence.
}
\description{
This function computes the Iterative Dressler-Shectman Test for a galaxy cluster and returns the number of iterations untill convergence.
}
\examples{
DresslerShectmanIter(cat)
}
