% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{DresslerShectmanTest2}
\alias{DresslerShectmanTest2}
\title{DresslerShectmanTest2}
\usage{
DresslerShectmanTest2(x, y, vel)
}
\arguments{
\item{cat}{data frame with a the angular coordinates (ra and dec in degrees) and the line of sight velocity (vel in km/s). The columns must be named 'ra', 'dec' and 'vel'}
}
\value{
The Delta value of the Dressler-Shectman test corresponding to the galaxy cluster.
}
\description{
This function returns the value of the Dressler-Shectman statistic for a galaxy cluster with ngal = sqrt(n).
}
\examples{
DresslerShectmanTest2(cat)
}
