% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{GalaxiesFeatures_new}
\alias{GalaxiesFeatures_new}
\title{GalaxiesFeatures_new}
\usage{
GalaxiesFeatures_new(group, featuresFunctions, featuresNames)
}
\arguments{
\item{group_gals}{Data frame with a the angular coordinates (ra and dec in radians), the redshift, the color (g-r), the apparent magnitude (r) and the id of the group for each galaxy. The columns must be named 'ra', 'dec' and 'z', 'color', 'mag' and 'id'.}

\item{group_data}{Data frame with a the data corresponding to the galaxy cluster. Must be an object with the same data as the output of ClusterFeatures.}
}
\value{
Data frame with the features of the galaxies.
}
\description{
This function estimates the features for the galaxies of a galaxy cluster.
}
\examples{
GalaxiesFeatures(group_gals, group_data)
}
