% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{get_clusters_classification}
\alias{get_clusters_classification}
\title{get_clusters_classification}
\usage{
get_clusters_classification(ClustersData, model)
}
\arguments{
\item{ClustersData}{Data frame with the features of all the galaxy clusters.}

\item{model}{Machine Learning model that predicts the dynamical status of a galaxy cluster. Can be the output of Train_cluster_model.}
}
\value{
Vector with the probabilities of being a merging cluster.
}
\description{
This function classify galaxies clusters.
}
\examples{
get_clusters_classification(ClustersData, model)
}
