% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{get_distance}
\alias{get_distance}
\title{get_distance}
\usage{
get_distance(x, x0, y, y0, ngal)
}
\arguments{
\item{x}{Vector with the Right ascensions in radians.}

\item{x0}{Right ascension of the central galaxy.}

\item{y}{Vector with the Declinations in radians.}

\item{y0}{Declination of the central galaxy.}

\item{ngal}{Number of galaxies to be returned}
}
\value{
Vector with the distances to the central galaxy.
}
\description{
This function returns the distance from ngal galaxies to a galaxy.
}
\examples{
get_distance(x, x0, y, y0, ngal)
}
