% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{get_galaxies_features}
\alias{get_galaxies_features}
\title{get_galaxies_features}
\usage{
get_galaxies_features(dat, ClustersData, name.gal)
}
\arguments{
\item{dat}{Data frame with a the angular coordinates (ra and dec in radians), the redshift, the color (g-r), the apparent magnitude (r) and the id of the group for each galaxy of each galaxy cluster. The columns must be named 'ra', 'dec' and 'z', 'color', 'mag' and 'id'.}

\item{ClustersData}{Data frame with a the information of the galaxy cluster. Must have the same data as the output of ClusterFeatures.}
}
\value{
Data frame with the features of all the galaxies of the catalog.
}
\description{
This function estimates the features for a catalog of galaxies.
}
\examples{
get_galaxies_features(dat, ClustersData)
}
