% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{get_substructures}
\alias{get_substructures}
\title{get_substructures}
\usage{
get_substructures(ClustersData, GalaxiesData, model)
}
\arguments{
\item{ClustersData}{Data frame with the features of all the galaxy clusters.}

\item{GalaxiesData}{Data frame with the features of all the galaxies of the clusters.}

\item{model}{Machine Learning model that predicts the probability of a galaxy of being part of a substructure inside a merging cluster. Can be the output of Train_galaxy_model.}
}
\value{
Data frame with the properties of the substructures.
}
\description{
This function look for the substructures inside merging clusters.
}
\examples{
get_substructures(ClustersData, GalaxiesData, model)
}
