% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{GroupFeatures}
\alias{GroupFeatures}
\title{groupFeatures}
\usage{
GroupFeatures(group_gals, group_data, featuresNames)
}
\arguments{
\item{group}{Data frame with the galaxies of the cluster.}
}
\value{
Vector with the properties of the substructures.
}
\description{
This function performs the identification of the substructures inside a galaxy cluster.
}
\examples{
SubstructureIdentification(group)
}
