% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Med}
\alias{Med}
\title{Simulated data with real mediation effect}
\format{
An object of class \code{list} of length 8.
}
\usage{
Med
}
\description{
Simulated data with real mediation effect
}
\examples{
data(Med)
}
\keyword{datasets}
