% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nullMed}
\alias{nullMed}
\title{Simulated data with null mediation effect}
\format{
An object of class \code{list} of length 4.
}
\usage{
nullMed
}
\description{
Simulated data with null mediation effect
}
\examples{
data(nullMed)
}
\keyword{datasets}
