% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{ProjectedDistance}
\alias{ProjectedDistance}
\title{ProjectedDistance}
\usage{
ProjectedDistance(cat)
}
\arguments{
\item{cat}{data frame with a the angular coordinates (ra and dec in radians) and the redshift. The columns must be named 'ra', 'dec' and 'z'.}
}
\value{
sum(1/d) where d is a vector with the projected distances.
}
\description{
This function estimates the projected distances between galaxies needed for the estimation of the virial radius of the galaxy cluster.
}
\examples{
ProjectedDistance(cat)
}
