% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{roc}
\alias{roc}
\title{roc}
\usage{
roc(dat, real.name = "class", pred.name = "merProb")
}
\arguments{
\item{dat}{Data frame with the real class and the prediction of the probability.}

\item{real.name}{String with the name of the variable that is the real class. Defatult is 'class'.}

\item{pred.name}{String with the name of the variable that is the predicted probability. Defaults is 'merProb'.}
}
\value{
data frame with the statistics.
}
\description{
This function estimates some statistics for measuring the performance of the machine learning model.
}
\examples{
roc(cat)
}
