% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{VelocityDispersionGAP}
\alias{VelocityDispersionGAP}
\title{VelocityDispersionGAP}
\usage{
VelocityDispersionGAP(x)
}
\arguments{
\item{x}{Vector with the velocities of the galaxies.}
}
\value{
The GAP velocity dispersion of the galaxy cluster.
}
\description{
This function estimates the velocity dispersion of a galaxy cluster with the GAP method.
}
\examples{
VelocityDispersionGAP(x)
}
