% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{DresslerShectmanTest2}
\alias{DresslerShectmanTest2}
\title{DresslerShectmanTest2}
\usage{
DresslerShectmanTest2(cat)
}
\arguments{
\item{cat}{data frame with a the angular coordinates (ra and dec in radians) and the redshift. The columns must be named 'ra', 'dec' and 'z'.}
}
\value{
The Delta value of the Dressler-Shectman test corresponding to the galaxy cluster.
}
\description{
This function returns the value of the Dressler-Shectman statistic for a galaxy cluster with ngal = sqrt(n).
}
\examples{
DresslerShectmanTest2(cat)
}
