% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesEstimationFunctions.R
\name{messi}
\alias{messi}
\title{messi}
\usage{
messi(cat = -99, clusters = TRUE, galaxies = TRUE,
  classification = TRUE, clustersOutput = "clustersOutput.dat",
  galaxiesOutput = "galaxiesOutput.dat",
  classOutput = "ClassOutput.dat", folder = "folder",
  ClustersML = "defaultClustersModel",
  GalaxiesML = "defaultGalaxiesModel", ntotal = 0, ngal.lim = 30,
  est = TRUE, probLimit = 0.3)
}
\arguments{
\item{cat}{Data frame with the catalog of galaxies. It must have the angular positions in radians (ra, dec), the redshift (z), a flag indicating to which clusters it belong (id), the r apparent magnitude (mag) and a the g-r color (color).}
}
\value{
NULL.
}
\description{
This function classify the clusters and estimates the merging substructures inside them.
}
\examples{
messi(cat)
}
